* commit '1bf27d9dbe258ab19abb2ba3aa4361f50da8b93d': (54 commits)
  IFS-6696 Update AT's
  IFS-6696 - Visual cleanup to unit tests.
  IFS-6696 Update AT's
  IFS-6696 fix js problem
  IFS-6696 Update AT's
  IFS-6696 Update AT's
  IFS-6696 - Fix to failing documentation unit test.
  IFS-6696 Update AT's
  IFS-6696 - Commented code removed.
  IFS-6696 - Fix to unit test.
  IFS-6696 - Client-side errors updated.
  IFS-6696 Update AT's
  IFS-6696 - Project setup yml updated with toggle name and value, non-number data prevent input added to spend profile and funding amount of other funding.
  IFS-6696 - Check for is requesting funding added to data service returning validation error when requesting funding is false.
  IFS-6696 - ISE in your funding for invited PS partner fixed.
  IFS-6696 fixd unit test
  IFS-6696 Update AT's
  IFS-6696 - Fix to compareTo percentage to max in data-service.
  IFS-6696 - ISE resolved in your organisation.
  IFS-6696 - Migration script remove update.
  ...

+++ ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/testdata/services/ApplicationDataBuilderService.java
@@ -393,7 +393,7 @@ public class ApplicationDataBuilderService extends BaseDataBuilderService {
             case "Working days per year":
                 return builder.withWorkingDaysPerYear(Integer.valueOf(financeRow.metadata.get(0)));
             case "Grant claim":
-                return builder.withGrantClaim(Integer.valueOf(financeRow.metadata.get(0)));
+                return builder.withGrantClaim(BigDecimal.valueOf(Integer.valueOf(financeRow.metadata.get(0))));
             case "Organisation size":
                 return builder.withOrganisationSize(OrganisationSize.findById(Long.valueOf(financeRow.metadata.get(0))));
             case "Work postcode":
@@ -525,7 +525,7 @@ public class ApplicationDataBuilderService extends BaseDataBuilderService {
                         builder[0] = builder[0].withVat(true);
                         break;
                     case FINANCE:
-                        builder[0] = builder[0].withGrantClaim(30);
+                        builder[0] = builder[0].withGrantClaim(BigDecimal.valueOf(30));
                         break;
                     case GRANT_CLAIM_AMOUNT:
                         builder[0] = builder[0].withGrantClaimAmount(12000);
@@ -575,7 +575,7 @@ public class ApplicationDataBuilderService extends BaseDataBuilderService {
                 withUser(user).
                 withAcademicCosts(costs -> costs.
                         withTsbReference("My REF").
-                        withGrantClaim(100).
+                        withGrantClaim(BigDecimal.valueOf(100)).
                         withOtherFunding("Lottery", LocalDate.of(2016, 4, 1), bd("2468")).
                         withDirectlyIncurredStaff(bd("22")).
                         withDirectlyIncurredTravelAndSubsistence(bd("44")).
